Remove Founders&Coders from Beginners section
Given our conversation about keeping meetups for underrepresented groups in that category and nowhere else, I've deleted the duplicate founders & coders from the beginners section and left the one that was in the u/r groups section. I hope that's ok! Let me know if not? Thank you soo much!