Alphabetical order
Putting the beginners and u/r group sections into alphabetical order.
Please note that this is a kind of approximate alphabetical order (!) for example "London App Brewery" comes first in the beginners section, because I've ignored the word London (given this is present in so many meetup names) and used the word "App" to sort into alphabetical order. If there are any problems with my method of sorting, please just let me know and I will be more than happy to change! Thank you!